requirements.txt
# nazBot v4.0 — Python Dependencies
# Install: pip install -r requirements.txt

# Binance Futures API
python-binance==1.0.19

# Technical Analysis
pandas>=1.5.0
ta>=0.10.2
numpy>=1.24.0

# Web Dashboard
flask>=2.3.0
flask-cors>=4.0.0

# Async / Concurrency (standard, no extra install)
# threading, concurrent.futures → built-in

# Optional: for Replit keep-alive ping
requests>=2.31.0
